Fix incorrect copy path in install.sh
Line 4 (previously `sudo cp swi.vmware_vmnet.monitor.plist /Library/LaunchAgents/`) did not copy the .plist file to the correct location called by the launchctl load in the next line. Adding a ~ before the file path on this line puts the .plist file in the right location to be called in the launchctl command.